refactor(melee_phase): change the type of the global const value default_dice
default_dice is now a f64 as we mainly use it this way

+++ src/fight/global_values.rs
@@ -1,3 +1,3 @@
-pub const DEFAULT_DICE: usize = 6;
+pub const DEFAULT_DICE: f64 = 6.0_f64;
 /// The key value to compute the minimum roll when attempting a save
 pub const ARMOUR_SAVE_THRESHOLD: usize = 7;
